refactor(user-management): relocate byte-identical team/mfa models to autobot_shared (#12647)
